backend/requirements.txt
python-dotenv>=0.15.0,<1.0
pymongo>=3.11,<4.0
requests>=2.25,<3.0
pandas>=1.2,<2.0
plotly>=4.14,<5.0
numpy>=1.20,<2.0
pyyaml>=5.4,<6.0
ta>=0.7,<1.0
fastapi>=0.63,<1.0
uvicorn>=0.13.4,<1.0